feat: work on refs
