+++ src/pages/forms/checkboxes/checkboxes.tsx
@@ -1,17 +1,6 @@
 import React, { Component } from 'react';
+import { CHECKBOXES_MOCKS } from 'src/utils/const/texts';
 
-export const CHECKBOXES_MOCKS = [
-  {
-    name: 'email',
-    text: 'Send me emails',
-    defaultChecked: true,
-  },
-  {
-    name: 'sms',
-    text: 'Send me sms',
-    defaultChecked: false,
-  },
-];
 export default class CheckboxesInputs extends Component {
   render() {
     return (
